fix javac [cast] warnings

+++ src/test/loops/FJPhaserJacobi.java
@@ -142,8 +142,8 @@ public class FJPhaserJacobi {
         public void compute() {
             int rows = hiRow - loRow + 1;
             int cols = hiCol - loCol + 1;
-            int rblocks = (int)(Math.round((float)rows / dimGran));
-            int cblocks = (int)(Math.round((float)cols / dimGran));
+            int rblocks = Math.round((float)rows / dimGran);
+            int cblocks = Math.round((float)cols / dimGran);
 
             int n = rblocks * cblocks;
